fix(ci): run rust job from backend/ and clear the clippy -D warnings gate
The rust CI job ran cargo build/test/clippy from the repo root, but there is
no root Cargo.toml — the backend crate lives in backend/ — so every cargo step
failed at "could not find Cargo.toml". Add a job-level
`working-directory: backend` and point rust-cache at `workspaces: backend`.

With the job actually reachable, `cargo clippy -- -D warnings` surfaced 26
lint failures. Cleared them all:
  * Mechanical (clippy --fix): redundant `as i32` no-op casts, needless
    closures/borrows, `.await.ok()` -> `if let Ok`, manual range checks ->
    RangeInclusive::contains, derived Default, multi-char replace.
  * logging: collapse the identical 3xx/4xx WARN branches into one.
  * ws/game: clamp() instead of .max().min() on player coords.
  * developer: demote orphaned doc comments (structs long removed) to //.
  * Deliberate patterns annotated with #[allow]: NotificationType::from_str
    (should_implement_trait), too-many-args on three query builders,
    type_complexity on load_receipt.

Verified by running from backend/: `cargo clippy -- -D warnings` (clean) and
`cargo test` (all suites pass, 0 failed; live_db_tests remain #[ignore]d).

+++ backend/src/api/games.rs
@@ -195,7 +195,7 @@ pub async fn get_leaderboard(
     Path(game_id): Path<Uuid>,
     Query(query): Query<LeaderboardQuery>,
 ) -> Result<Json<crate::api::response::PaginatedResponse<LeaderboardEntry>>> {
-    let limit = query.limit.unwrap_or(100) as i32;
+    let limit = query.limit.unwrap_or(100);
 
     let entries = sqlx::query_as::<_, (Uuid, String, i64)>(
         "SELECT u.id, u.username, ghs.score
